analyzer/emerald: rename emerald->runtime
