+++ analyzer/runtime/incoming.go
@@ -1,4 +1,4 @@
-package emerald
+package runtime
 
 import (
 	"bytes"
