+++ analyzer/runtime/runtime.go
@@ -1,4 +1,4 @@
-package emerald
+package runtime
 
 import (
 	"context"
