+++ cmd/analyzer/analyzer.go
@@ -15,8 +15,8 @@ import (
 
 	"github.com/oasisprotocol/oasis-indexer/analyzer"
 	"github.com/oasisprotocol/oasis-indexer/analyzer/consensus"
-	"github.com/oasisprotocol/oasis-indexer/analyzer/emerald"
 	"github.com/oasisprotocol/oasis-indexer/analyzer/evmtokens"
+	"github.com/oasisprotocol/oasis-indexer/analyzer/runtime"
 	"github.com/oasisprotocol/oasis-indexer/cmd/common"
 	"github.com/oasisprotocol/oasis-indexer/config"
 	"github.com/oasisprotocol/oasis-indexer/log"
@@ -157,8 +157,8 @@ func NewService(cfg *config.AnalysisConfig) (*Service, error) {
 		switch analyzerCfg.Name {
 		case consensus.ConsensusDamaskAnalyzerName, "consensus_main_damask": // TODO: drop "main" variant; as of Oct 2022, it exists only to support legacy helmfiles
 			a, err = consensus.NewMain(analyzerCfg, client, logger)
-		case emerald.EmeraldDamaskAnalyzerName, "emerald_main_damask": // TODO: drop "main" variant; as of Oct 2022, it exists only to support legacy helmfiles
-			a, err = emerald.NewMain(analyzerCfg, client, logger)
+		case runtime.EmeraldDamaskAnalyzerName, "emerald_main_damask": // TODO: drop "main" variant; as of Oct 2022, it exists only to support legacy helmfiles
+			a, err = runtime.NewMain(analyzerCfg, client, logger)
 		case evmtokens.EmeraldDamaskTokensAnalyzerName:
 			a, err = evmtokens.NewMain(analyzerCfg, client, logger)
 		case analyzer.MetadataRegistryAnalyzerName:
